+++ packages/react-magma-dom/src/components/Breadcrumb/BreadcrumbItem.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
-import styled from '../../theme/styled';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { Hyperlink } from '../Hyperlink';
 import { useIsInverse } from '../../inverse';
 import { ChevronRightIcon } from 'react-magma-icons';
+import styled, { CreateStyled } from '@emotion/styled';
+import { ThemeInterface } from '../../theme/magma';
 
 /**
  * @children required
@@ -20,6 +21,8 @@ export interface BreadcrumbItemProps
   to?: string;
 }
 
+const typedStyled = styled as CreateStyled<ThemeInterface>;
+
 const StyledItem = styled.li`
   list-style: none;
   margin: 0;
@@ -27,7 +30,7 @@ const StyledItem = styled.li`
   display: flex;
 `;
 
-const StyledSpan = styled.span<{ isInverse?: boolean }>`
+const StyledSpan = typedStyled.span<{ isInverse?: boolean }>`
   color: ${props =>
     props.isInverse
       ? props.theme.colors.neutral100