+++ packages/react-magma-dom/src/components/TimePicker/TimePicker.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import styled from '../../theme/styled';
 import { Announce } from '../Announce';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { AmPmToggle } from './AmPmToggle';
@@ -11,6 +10,8 @@ import { VisuallyHidden } from '../VisuallyHidden';
 import { useIsInverse } from '../../inverse';
 import { FormFieldContainer } from '../FormFieldContainer';
 import { inputWrapperStyles } from '../InputBase';
+import styled, { CreateStyled } from '@emotion/styled';
+import { ThemeInterface } from '../../theme/magma';
 
 export interface TimePickerProps extends UseTimePickerProps {
   /**
@@ -19,7 +20,9 @@ export interface TimePickerProps extends UseTimePickerProps {
   inputStyle?: React.CSSProperties;
 }
 
-const InputsContainer = styled.div<{
+const typedStyled = styled as CreateStyled<ThemeInterface>;
+
+const InputsContainer = typedStyled.div<{
   hasError?: boolean;
   isInverse?: boolean;
 }>`
@@ -37,7 +40,7 @@ const Divider = styled.span`
   top: -1px;
 `;
 
-const StyledNumInput = styled.input<{
+const StyledNumInput = typedStyled.input<{
   isInverse?: boolean;
 }>`
   border: 0;