+++ packages/react-magma-dom/src/components/Typography/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import styled from '../../theme/styled';
 import { css } from '@emotion/core';
+import styled from '@emotion/styled';
 
 export interface TypographyProps<T = HTMLParagraphElement>
   extends React.HTMLAttributes<T> {