+++ packages/charts/CHANGELOG.md
@@ -1,6 +1,36 @@
 # @react-magma/charts
 
-## 6.0.0
+## 7.0.0-next.4
+
+### Minor Changes
+
+- 88719d577: Update CarbonCharts broken style and docs
+
+## 7.0.0-next.3
+
+### Patch Changes
+
+- 8a5b4bedb: fix(CarbonCharts): Update styles
+
+## 7.0.0-next.2
+
+### Minor Changes
+
+- 1bb49bb0c: feat(CarbonCharts): Support Carbon Charts
+
+## 7.0.0-next.1
+
+### Patch Changes
+
+- Updated dependencies [ab6ffd4ed]
+  - react-magma-dom@4.4.0-next.2
+
+## 6.0.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [3e14d4bd8]
+  - react-magma-dom@4.3.1-next.0
 
 ## 6.0.0-next.1